pyproject.toml
[project]
name = "ai-evaluation-testing-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "openai>=2.14.0",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
